+++ test/suite/Matcher/Exception/UndefinedTypeExceptionTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Eloquent\Phony\Matcher\Exception;
+
+use PHPUnit\Framework\TestCase;
+
+class UndefinedTypeExceptionTest extends TestCase
+{
+    public function testException()
+    {
+        $type = Undefined::class;
+        $exception = new UndefinedTypeException($type);
+
+        $this->assertSame($type, $exception->type());
+        $this->assertSame(
+            "Undefined type 'Eloquent\\\\Phony\\\\Matcher\\\\Exception\\\\Undefined'.",
+            $exception->getMessage()
+        );
+        $this->assertSame(0, $exception->getCode());
+        $this->assertNull($exception->getPrevious());
+    }
+}